requirements.txt
mcp
docling
qdrant-client
openai
python-dotenv